added events parsing into packet, load from packet and client action visitor for events

+++ Client/PMStatistics.cpp
@@ -28,7 +28,7 @@ long long PMStatistics::packetRecieve(const int id)
 	recieved++;
 
 	const auto duration = chrono::duration_cast<chrono::microseconds>(t2 - times[id]).count();
-	latencyCalc.update(duration);
+	latencyCalc.update(static_cast<double>(duration));
 	lastLatency = duration;
 
 	times.erase(id);
